experimented with "this" java keyword
In this commit I have changed the listen function in the
OnClickListener class. I have removed the parameter taking an instance
of my class in favour of using the “this” keyword in java. “this”
refers to the current object in java. Since I am calling the listen
function from an instance of the OnClickListener class, I can use this
to refer to the object as oppose to entering the same object into the
function as a parameter.